New - koji_helpers::repo::latest parameter
This new parameter permits setting of the `latest` setting in the mash
configuration files, per repository.  The default value matches that of
the underlying mash tool, as is common practice with this Puppet module,
which means if this is not set, behavior will not change from what is
was prior to its introduction here.

See also https://bugzilla.redhat.com/show_bug.cgi?id=1082830.